Project Deliverable 5/ApiDemoProj/requirements.txt
Django==1.10.6
googlemaps==2.4.6
oauthlib==2.0.2
PyJWT==1.5.0
python-google-places==1.4.0
python-openid==2.2.5
requests==2.11.1
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
